backend/requirements.txt
Django==4.0.4
django-environ==0.9.0
psycopg2-binary==2.9.3
djangorestframework==3.13.1
django-filter==21.1
django-cors-headers==3.13.0
Pillow==9.1.1
django-cleanup==6.0.0
google-api-python-client==2.51.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
oauth2client